Fix start and end time in analysis
We were using the start of the first thread-run as the initial event, but
thread lifetimes start before it gets run. We now use the Startup event as the
initial timestamp, and the last CapsetDelete as the final timestamp (there is a
Shutdown event but it does not seem to be generated.