Make sure _shutdown is always set
Very often event logs don't contain a CapDelete or Shutdown event (even
when you exit the program properly). When this is the case
ghc-events-analyze crashes with the message "_shutdown not set". This is
annoying since the whole event log could in principle be visualized.

This patch just marks the last event time as the shutdown time.