Fix compilation for newer GHCs:
Now compiles with 8.0.2, 8.2.2 and 8.4.2. When compiling with 7.10.3 I
get this error:

    Building library for zlib-0.6.2..

    [1 of 5] Compiling Codec.Compression.Zlib.Stream (
    dist/build/Codec/Co /Codec/Compression/Zlib/Stream.o )
    /usr/bin/x86_64-linux-gnu-ld: -r and -pie may not be used together

    collect2: error: ld returned 1 exit status

    cabal: Failed to build zlib-0.6.2 (which is required by
    exe:ghc-events from ghc-events-analyze-0.2.5). See the build log
    above for details.

I suspect this is due to the newer toolchain I have on my system. If
needed I can try to test this on an older system.

The important bit is `ghc-events` lower bound. Newer versions should be
backwards compatible but 0.6 introduced some breaking changes.